% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blame}
\alias{blame}
\title{Blame Data Set}
\format{
A data frame with these columns:
\describe{
\item{ptotpalest1}{The proportion of blame assigned to Palestinians}
\item{palestpk}{A binary variable indicating whether the Palestinians were listed as three groups or one, 0 = unpacked, 1 = packed}
}
}
\source{
Halevy, N., Maoz, I., Vani, P. and Reit, E.S., 2022. Where the blame lies: Unpacking groups into their constituent subgroups shifts judgments of blame in intergroup conflict. Psychological Science, 33(1), pp.76-89.
}
\usage{
blame
}
\description{
The data for this example are from a replication and extension of Halevy et al. (2022), the replication conducted by Smithson.  An Israeli adult sample was recruited using the Qualtrics panel platform during April 2021 (N = 462; mean age = 39.9, age SD = 15.5; 245 females, 214 males, 3 other).  The participants were asked to assign percentages of blame to Israelis and Palestinians for the unresolved Israeli-Palestinian conflict.  Two experimental conditions varied whether the Palestinians were listed as one population or as three separate movements, i.e., listed as Palestinians vs listed as The Palestinian Hamas movement, The Palestinian Fatah movement, The Palestinian Islamic Jihad movement.  The hypothesis to be tested here was that unpacking the Palestinians into three groups would increase the percentage of blame that the Israeli participants assigned to Palestinians.
}
\keyword{datasets}
