% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_wrappers.R
\name{ftcdfqMLE2}
\alias{ftcdfqMLE2}
\alias{print.ftcdfq2}
\alias{summary.ftcdfq2}
\alias{print.summary.ftcdfq2}
\alias{coef.ftcdfq2}
\alias{logLik.ftcdfq2}
\title{Maximum Likelihood Estimation for 2-Parameter CDF Quantile Regression}
\usage{
ftcdfqMLE2(formula, data, start = NULL)

\method{print}{ftcdfq2}(x, ...)

\method{summary}{ftcdfq2}(object, ...)

\method{print}{summary.ftcdfq2}(x, ...)

\method{coef}{ftcdfq2}(object, ...)

\method{logLik}{ftcdfq2}(object, ...)
}
\arguments{
\item{formula}{A two-part formula separated by pipes (\code{|}),
where the first part specifies predictors
for mu (location) and the second for sigma (dispersion).}

\item{data}{A data frame containing the variables specified in the formula.}

\item{start}{An optional numeric vector of starting values for the parameters.
If \code{NULL}, defaults to \code{0.1} for all parameters.}

\item{x}{An object of class \code{"ftcdfq2"} or \code{"summary.ftcdfq2"} (for print methods).}

\item{...}{Additional arguments (not used).}

\item{object}{An object of class \code{"ftcdfq2"}.}
}
\value{
An object of class (\code{"ftcdfq2"}.) This is a list containing:
\describe{
}

An object of class \code{"summary.ftcdfq2"}.
}
\description{
Fits a parametric frequentist regression model for bounded outcomes on the unit interval
using maximum likelihood estimation across two distribution parameters: mu, sigma.
}
\examples{
\dontrun{
# Select a distribution first
select_ftcdfq2()

# Fit an MLE model
m1 <- ftcdfqMLE2(ptotpalest1 ~ palestpk | 1 | 1 , data = blame)
summary(m1)
}
# Fit a model
}
