% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_wrappers.R
\name{parameter_cor2}
\alias{parameter_cor2}
\title{Extract Parameter Correlation Matrix for ftcdfq2 Fits}
\usage{
parameter_cor2(object, ...)
}
\arguments{
\item{object}{An object of class \code{"exftcdfq"}.}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric correlation matrix.
}
\description{
Calculates the correlation matrix of the estimated parameters from a fitted model.
}
