% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_wrappers.R
\name{predict.ftcdfq2}
\alias{predict.ftcdfq2}
\title{Predicting quantiles
Obtains predicted values (quantiles) from a fitted CDF quantile regression model object.}
\usage{
\method{predict}{ftcdfq2}(object, newdata = NULL, q = 0.5, ...)
}
\arguments{
\item{object}{An object of class \(\code{"ftcdfq2"}.\)}

\item{newdata}{An optional data frame or vector of new values for out-of-sample prediction.
If omitted, fitted values for the original dataset are returned.}

\item{q}{A numeric value or vector indicating the target quantile(s). Defaults to \code{0.5} (median).}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric vector of predicted values.
}
\description{
Predicting quantiles
Obtains predicted values (quantiles) from a fitted CDF quantile regression model object.
}
