% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_wrappers.R
\name{residuals.exftcdfq}
\alias{residuals.exftcdfq}
\title{Extract Model Residuals}
\usage{
\method{residuals}{exftcdfq}(object, ...)
}
\arguments{
\item{object}{An object of class \(\code{"exftcdfq"}.\)}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric vector of residuals.
}
\description{
Calculates the raw residuals (Observed minus Predicted Medians) for an exftcdfq object.
}
