% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_wrappers.R
\name{residuals.ftcdfq3}
\alias{residuals.ftcdfq3}
\title{Extract Model Residuals}
\usage{
\method{residuals}{ftcdfq3}(object, ...)
}
\arguments{
\item{object}{An object of class \(\code{"ftcdfq3"}.\)}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric vector of residuals.
}
\description{
Calculates the raw residuals (Observed minus Predicted Medians) for a ftcdfq3 object.
}
