% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yoon}
\alias{yoon}
\title{Yoon Data Set}
\format{
A data frame with these columns:
\describe{
\item{pphysdecis}{The proportion of total time spent by each patient in the ED under assessment and treatment by a physician}
\item{Ambulance}{A binary variable indicating how each patient arrived at the ED, 0 = walk-in, 1 = ambulance arrival}
}
}
\source{
Yoon, P., I. Steiner, and G. Reinhardt. 2003. Analysis of factors influencing length of stay in the emergency department. CJEM 5 (3):155–61.
}
\usage{
yoon
}
\description{
The data are records of 894 patients whose times were recorded for five stages of ED assessment and treatment from a study consisting of recorded information on patients admitted to the emergency department of the University of Alberta Hospital between midnight January 23 and midnight January 29, 1999 (Yoon, Steiner, and Reinhardt 2003).
}
\keyword{datasets}
