% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_wrappers.R
\name{exftcdfq}
\alias{exftcdfq}
\title{Fit an EXFTCDFQ Model using brms}
\usage{
exftcdfq(formula, data, backend = "cmdstanr", ...)
}
\arguments{
\item{formula}{A brms formula object.}

\item{data}{A data frame containing the variables in the model.}

\item{backend}{The backend engine to use for compilation (defaults to "cmdstanr").}

\item{...}{Additional arguments passed to brms::brm.}
}
\description{
Fit an EXFTCDFQ Model using brms
}
